backend/requirements/dev.txt
Django==5.2.9
python-decouple==3.8
redis==7.3.0
django-cors-headers==4.9.0
django-modeltranslation==0.19.17
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-rosetta==0.10.3
drf-yasg==1.21.11
drf-spectacular==0.29.0
psycopg2-binary==2.9.11
martor==1.7.16
pillow==12.1.1

django-debug-toolbar==6.2.0

channels==4.2.0
channels-redis==4.2.1
uvicorn[standard]==0.34.0